Styling overhaul of the inbox page with code quality changes
