+++ package.json
@@ -22,6 +22,7 @@
   "homepage": "https://github.com/Intro-to-SE-Lab-Fall-20/Group-12#readme",
   "dependencies": {
     "@hapi/joi": "^17.1.1",
+    "dayjs": "^1.9.3",
     "dotenv": "^8.2.0",
     "ejs": "^3.1.5",
     "express": "^4.17.1",
